+++ Terra/Changes.md
@@ -1,7 +1,9 @@
 # Utopia Terra - list of changes
 
 ## v1.2.1 (in development)
-Adding hidden height calculations
+### New features
+- Added `.calculateHiddenHeight(...)` function to **GlobeMath**, 
+  which calculates how much of a viewed object should be hidden behind a physical horizon.
 
 ## v1.2 - 04.10.2024
 Made it easier to support more generic world views by introducing a couple new traits.